+++ server/src/main/java/com/inventage/portal/gateway/proxy/middleware/csp/compositeCSP/CompositeCSPHandler.java
@@ -1,0 +1,9 @@
+package com.inventage.portal.gateway.proxy.middleware.csp.compositeCSP;
+
+import io.vertx.ext.web.handler.CSPHandler;
+
+public interface CompositeCSPHandler extends CSPHandler {
+    static CSPHandler create() {
+        return new CompositeCSPHandlerImpl();
+    }
+}