+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Middleware `claimToHeader` for setting an HTTP header from a JWT claim value ([PORTAL-1483](https://issue.inventage.com/browse/PORTAL-1483)).
 - Support for additional issuers. Can be defined in the configuration with `additionalIssuers` (([PORTAL-1331](https://issue.inventage.com/browse/PORTAL-1331)).
+- Configuring the CSP middleware at route level. It is now possible to define basic CSP policies on the entry-middleware and specific/more restrictive CSP policies on each specific route (([PORTAL-1230](https://issue.inventage.com/browse/PORTAL-1230)).
 
 ## [8.0.0]-202305240847-789-bdd58cd9 - 2023-05-24
 - Created middleware `bodyHandler`, that is required for the csrf middleware ([PORTAL-1497](https://issue.inventage.com/browse/PORTAL-1497)).